D0001: Hardware Build
The repository has been linked to the TIA Portal
Workspace for the PS2001-pal-software module build.

This build is the TIA Project HARDWARE BUILD only

The CPU is a 1515-2PN (6ES7 515-2AM02-0AB0) SW Ver 2.8

The CPU is designated CON100 and has IP address:

   192.168.1.100

Consequently, the Workspace repository for the TIA Project
now has a corresponding CON100 subfolder this is equipped with:

1. Program blocks
   * OB00001_IntNMainProgram.xml

2. PLC tags
   * Default tag table.xml

The default tag table, is that produce when the project
was created.

OB1 has been equipped with standard comments and the
MIT Licence. It contains no code at this point.