D0002: Baseline Build
This is the formal BASELINE build of the PAL software.

The baseline build can be used to import any of the
PAL software versions. It is equipped with the following:

### Program blocks
 - OB00001_IntNMainProgram
 - FC61000_DocGenExample

### PLC data types
 - UT01000_St_SysRevision
 - UT01001_St_SysLicence

### PLC tags
 - Default tag table
 - PAL_SystemTags

Those aspects of the PAL project that have been migrated to
the GitHub Version Control System (VCS) are operating under
the Software Revision Mechanism (SCM) specified in document
PS2001-5-2302-011, this is available here:

   https://practicalseries.com/2001-pal/31-git/11-00-scm.html